// src/types/legend.ts
interface Legend {
    birthday: string
    blendingInstanceUsed: string
    id: string
    image: string
    isDestroyed: boolean
    isHatched: boolean
    isLegendary: boolean
    legendCreator: string
    parent1: string
    parent2: string
    postfix: string
    prefix: string 
    season: string
    totalOffspring: number
}

export { type Legend }